+++ crates/control/src/button_filter.rs
@@ -17,8 +17,8 @@ pub struct ButtonFilter {
 #[context]
 pub struct CreationContext {
     pub calibration_buttons_timeout:
-        Parameter<Duration, "control.button_filter.calibration_buttons_timeout">,
-    pub head_buttons_timeout: Parameter<Duration, "control.button_filter.head_buttons_timeout">,
+        Parameter<Duration, "button_filter.calibration_buttons_timeout">,
+    pub head_buttons_timeout: Parameter<Duration, "button_filter.head_buttons_timeout">,
 }
 
 #[context]
@@ -27,8 +27,8 @@ pub struct CycleContext {
     pub cycle_time: Input<CycleTime, "cycle_time">,
 
     pub calibration_buttons_timeout:
-        Parameter<Duration, "control.button_filter.calibration_buttons_timeout">,
-    pub head_buttons_timeout: Parameter<Duration, "control.button_filter.head_buttons_timeout">,
+        Parameter<Duration, "button_filter.calibration_buttons_timeout">,
+    pub head_buttons_timeout: Parameter<Duration, "button_filter.head_buttons_timeout">,
 }
 
 #[context]
