+++ crates/control/src/fall_state_estimation.rs
@@ -18,8 +18,7 @@ pub struct FallStateEstimation {
 
 #[context]
 pub struct CreationContext {
-    pub fall_state_estimation:
-        Parameter<FallStateEstimationConfiguration, "control.fall_state_estimation">,
+    pub fall_state_estimation: Parameter<FallStateEstimationConfiguration, "fall_state_estimation">,
 }
 
 #[context]
@@ -32,8 +31,7 @@ pub struct CycleContext {
     pub filtered_roll_pitch: AdditionalOutput<Vector2<f32>, "filtered_roll_pitch">,
     pub forward_gravitational_difference: AdditionalOutput<f32, "forward_gravitational_difference">,
 
-    pub fall_state_estimation:
-        Parameter<FallStateEstimationConfiguration, "control.fall_state_estimation">,
+    pub fall_state_estimation: Parameter<FallStateEstimationConfiguration, "fall_state_estimation">,
 
     pub has_ground_contact: Input<bool, "has_ground_contact">,
     pub sensor_data: Input<SensorData, "sensor_data">,
