+++ crates/control/src/ground_contact_detector.rs
@@ -14,9 +14,9 @@ pub struct GroundContactDetector {
 
 #[context]
 pub struct CreationContext {
-    pub hysteresis: Parameter<f32, "control.ground_contact_detector.hysteresis">,
-    pub pressure_threshold: Parameter<f32, "control.ground_contact_detector.pressure_threshold">,
-    pub timeout: Parameter<Duration, "control.ground_contact_detector.timeout">,
+    pub hysteresis: Parameter<f32, "ground_contact_detector.hysteresis">,
+    pub pressure_threshold: Parameter<f32, "ground_contact_detector.pressure_threshold">,
+    pub timeout: Parameter<Duration, "ground_contact_detector.timeout">,
 }
 
 #[context]
@@ -25,9 +25,9 @@ pub struct CycleContext {
     pub sole_pressure: Input<SolePressure, "sole_pressure">,
     pub cycle_time: Input<CycleTime, "cycle_time">,
 
-    pub hysteresis: Parameter<f32, "control.ground_contact_detector.hysteresis">,
-    pub pressure_threshold: Parameter<f32, "control.ground_contact_detector.pressure_threshold">,
-    pub timeout: Parameter<Duration, "control.ground_contact_detector.timeout">,
+    pub hysteresis: Parameter<f32, "ground_contact_detector.hysteresis">,
+    pub pressure_threshold: Parameter<f32, "ground_contact_detector.pressure_threshold">,
+    pub timeout: Parameter<Duration, "ground_contact_detector.timeout">,
 }
 
 #[context]
