+++ crates/control/src/limb_projector.rs
@@ -8,16 +8,14 @@ pub struct LimbProjector {}
 
 #[context]
 pub struct CreationContext {
-    pub foot_bounding_polygon:
-        Parameter<Vec<Point3<f32>>, "control.projected_limbs.foot_bounding_polygon">,
-    pub knee_bounding_polygon:
-        Parameter<Vec<Point3<f32>>, "control.projected_limbs.knee_bounding_polygon">,
+    pub foot_bounding_polygon: Parameter<Vec<Point3<f32>>, "projected_limbs.foot_bounding_polygon">,
+    pub knee_bounding_polygon: Parameter<Vec<Point3<f32>>, "projected_limbs.knee_bounding_polygon">,
     pub lower_arm_bounding_polygon:
-        Parameter<Vec<Point3<f32>>, "control.projected_limbs.lower_arm_bounding_polygon">,
+        Parameter<Vec<Point3<f32>>, "projected_limbs.lower_arm_bounding_polygon">,
     pub torso_bounding_polygon:
-        Parameter<Vec<Point3<f32>>, "control.projected_limbs.torso_bounding_polygon">,
+        Parameter<Vec<Point3<f32>>, "projected_limbs.torso_bounding_polygon">,
     pub upper_arm_bounding_polygon:
-        Parameter<Vec<Point3<f32>>, "control.projected_limbs.upper_arm_bounding_polygon">,
+        Parameter<Vec<Point3<f32>>, "projected_limbs.upper_arm_bounding_polygon">,
 }
 
 #[context]
@@ -25,16 +23,14 @@ pub struct CycleContext {
     pub camera_matrices: RequiredInput<Option<CameraMatrices>, "camera_matrices?">,
     pub robot_kinematics: Input<RobotKinematics, "robot_kinematics">,
 
-    pub foot_bounding_polygon:
-        Parameter<Vec<Point3<f32>>, "control.projected_limbs.foot_bounding_polygon">,
-    pub knee_bounding_polygon:
-        Parameter<Vec<Point3<f32>>, "control.projected_limbs.knee_bounding_polygon">,
+    pub foot_bounding_polygon: Parameter<Vec<Point3<f32>>, "projected_limbs.foot_bounding_polygon">,
+    pub knee_bounding_polygon: Parameter<Vec<Point3<f32>>, "projected_limbs.knee_bounding_polygon">,
     pub lower_arm_bounding_polygon:
-        Parameter<Vec<Point3<f32>>, "control.projected_limbs.lower_arm_bounding_polygon">,
+        Parameter<Vec<Point3<f32>>, "projected_limbs.lower_arm_bounding_polygon">,
     pub torso_bounding_polygon:
-        Parameter<Vec<Point3<f32>>, "control.projected_limbs.torso_bounding_polygon">,
+        Parameter<Vec<Point3<f32>>, "projected_limbs.torso_bounding_polygon">,
     pub upper_arm_bounding_polygon:
-        Parameter<Vec<Point3<f32>>, "control.projected_limbs.upper_arm_bounding_polygon">,
+        Parameter<Vec<Point3<f32>>, "projected_limbs.upper_arm_bounding_polygon">,
 }
 
 #[context]
