+++ crates/control/src/localization.rs
@@ -28,36 +28,32 @@ pub struct Localization {
 
 #[context]
 pub struct CreationContext {
-    pub circle_measurement_noise:
-        Parameter<Vector2<f32>, "control.localization.circle_measurement_noise">,
+    pub circle_measurement_noise: Parameter<Vector2<f32>, "localization.circle_measurement_noise">,
     pub field_dimensions: Parameter<FieldDimensions, "field_dimensions">,
-    pub good_matching_threshold: Parameter<f32, "control.localization.good_matching_threshold">,
+    pub good_matching_threshold: Parameter<f32, "localization.good_matching_threshold">,
     pub gradient_convergence_threshold:
-        Parameter<f32, "control.localization.gradient_convergence_threshold">,
-    pub gradient_descent_step_size:
-        Parameter<f32, "control.localization.gradient_descent_step_size">,
+        Parameter<f32, "localization.gradient_convergence_threshold">,
+    pub gradient_descent_step_size: Parameter<f32, "localization.gradient_descent_step_size">,
     pub hypothesis_prediction_score_reduction_factor:
-        Parameter<f32, "control.localization.hypothesis_prediction_score_reduction_factor">,
-    pub hypothesis_retain_factor: Parameter<f32, "control.localization.hypothesis_retain_factor">,
+        Parameter<f32, "localization.hypothesis_prediction_score_reduction_factor">,
+    pub hypothesis_retain_factor: Parameter<f32, "localization.hypothesis_retain_factor">,
     pub hypothesis_score_base_increase:
-        Parameter<f32, "control.localization.hypothesis_score_base_increase">,
+        Parameter<f32, "localization.hypothesis_score_base_increase">,
     pub initial_hypothesis_covariance:
-        Parameter<Matrix3<f32>, "control.localization.initial_hypothesis_covariance">,
-    pub initial_hypothesis_score: Parameter<f32, "control.localization.initial_hypothesis_score">,
-    pub initial_poses: Parameter<Players<InitialPose>, "control.localization.initial_poses">,
-    pub line_length_acceptance_factor:
-        Parameter<f32, "control.localization.line_length_acceptance_factor">,
-    pub line_measurement_noise:
-        Parameter<Vector2<f32>, "control.localization.line_measurement_noise">,
+        Parameter<Matrix3<f32>, "localization.initial_hypothesis_covariance">,
+    pub initial_hypothesis_score: Parameter<f32, "localization.initial_hypothesis_score">,
+    pub initial_poses: Parameter<Players<InitialPose>, "localization.initial_poses">,
+    pub line_length_acceptance_factor: Parameter<f32, "localization.line_length_acceptance_factor">,
+    pub line_measurement_noise: Parameter<Vector2<f32>, "localization.line_measurement_noise">,
     pub maximum_amount_of_gradient_descent_iterations:
-        Parameter<usize, "control.localization.maximum_amount_of_gradient_descent_iterations">,
+        Parameter<usize, "localization.maximum_amount_of_gradient_descent_iterations">,
     pub maximum_amount_of_outer_iterations:
-        Parameter<usize, "control.localization.maximum_amount_of_outer_iterations">,
-    pub minimum_fit_error: Parameter<f32, "control.localization.minimum_fit_error">,
-    pub odometry_noise: Parameter<Vector3<f32>, "control.localization.odometry_noise">,
+        Parameter<usize, "localization.maximum_amount_of_outer_iterations">,
+    pub minimum_fit_error: Parameter<f32, "localization.minimum_fit_error">,
+    pub odometry_noise: Parameter<Vector3<f32>, "localization.odometry_noise">,
     pub player_number: Parameter<PlayerNumber, "player_number">,
-    pub score_per_good_match: Parameter<f32, "control.localization.score_per_good_match">,
-    pub use_line_measurements: Parameter<bool, "control.localization.use_line_measurements">,
+    pub score_per_good_match: Parameter<f32, "localization.score_per_good_match">,
+    pub use_line_measurements: Parameter<bool, "localization.use_line_measurements">,
 
     pub robot_to_field: PersistentState<Isometry2<f32>, "robot_to_field">,
 }
@@ -78,36 +74,32 @@ pub struct CycleContext {
     pub has_ground_contact: Input<bool, "has_ground_contact">,
     pub primary_state: Input<PrimaryState, "primary_state">,
 
-    pub circle_measurement_noise:
-        Parameter<Vector2<f32>, "control.localization.circle_measurement_noise">,
+    pub circle_measurement_noise: Parameter<Vector2<f32>, "localization.circle_measurement_noise">,
     pub field_dimensions: Parameter<FieldDimensions, "field_dimensions">,
-    pub good_matching_threshold: Parameter<f32, "control.localization.good_matching_threshold">,
+    pub good_matching_threshold: Parameter<f32, "localization.good_matching_threshold">,
     pub gradient_convergence_threshold:
-        Parameter<f32, "control.localization.gradient_convergence_threshold">,
-    pub gradient_descent_step_size:
-        Parameter<f32, "control.localization.gradient_descent_step_size">,
+        Parameter<f32, "localization.gradient_convergence_threshold">,
+    pub gradient_descent_step_size: Parameter<f32, "localization.gradient_descent_step_size">,
     pub hypothesis_prediction_score_reduction_factor:
-        Parameter<f32, "control.localization.hypothesis_prediction_score_reduction_factor">,
-    pub hypothesis_retain_factor: Parameter<f32, "control.localization.hypothesis_retain_factor">,
+        Parameter<f32, "localization.hypothesis_prediction_score_reduction_factor">,
+    pub hypothesis_retain_factor: Parameter<f32, "localization.hypothesis_retain_factor">,
     pub hypothesis_score_base_increase:
-        Parameter<f32, "control.localization.hypothesis_score_base_increase">,
+        Parameter<f32, "localization.hypothesis_score_base_increase">,
     pub initial_hypothesis_covariance:
-        Parameter<Matrix3<f32>, "control.localization.initial_hypothesis_covariance">,
-    pub initial_hypothesis_score: Parameter<f32, "control.localization.initial_hypothesis_score">,
-    pub initial_poses: Parameter<Players<InitialPose>, "control.localization.initial_poses">,
-    pub line_length_acceptance_factor:
-        Parameter<f32, "control.localization.line_length_acceptance_factor">,
-    pub line_measurement_noise:
-        Parameter<Vector2<f32>, "control.localization.line_measurement_noise">,
+        Parameter<Matrix3<f32>, "localization.initial_hypothesis_covariance">,
+    pub initial_hypothesis_score: Parameter<f32, "localization.initial_hypothesis_score">,
+    pub initial_poses: Parameter<Players<InitialPose>, "localization.initial_poses">,
+    pub line_length_acceptance_factor: Parameter<f32, "localization.line_length_acceptance_factor">,
+    pub line_measurement_noise: Parameter<Vector2<f32>, "localization.line_measurement_noise">,
     pub maximum_amount_of_gradient_descent_iterations:
-        Parameter<usize, "control.localization.maximum_amount_of_gradient_descent_iterations">,
+        Parameter<usize, "localization.maximum_amount_of_gradient_descent_iterations">,
     pub maximum_amount_of_outer_iterations:
-        Parameter<usize, "control.localization.maximum_amount_of_outer_iterations">,
-    pub minimum_fit_error: Parameter<f32, "control.localization.minimum_fit_error">,
-    pub odometry_noise: Parameter<Vector3<f32>, "control.localization.odometry_noise">,
+        Parameter<usize, "localization.maximum_amount_of_outer_iterations">,
+    pub minimum_fit_error: Parameter<f32, "localization.minimum_fit_error">,
+    pub odometry_noise: Parameter<Vector3<f32>, "localization.odometry_noise">,
     pub player_number: Parameter<PlayerNumber, "player_number">,
-    pub score_per_good_match: Parameter<f32, "control.localization.score_per_good_match">,
-    pub use_line_measurements: Parameter<bool, "control.localization.use_line_measurements">,
+    pub score_per_good_match: Parameter<f32, "localization.score_per_good_match">,
+    pub use_line_measurements: Parameter<bool, "localization.use_line_measurements">,
 
     pub line_data_bottom: PerceptionInput<Option<LineData>, "VisionBottom", "line_data?">,
     pub line_data_top: PerceptionInput<Option<LineData>, "VisionTop", "line_data?">,
