+++ crates/control/src/motion/head_motion.rs
@@ -13,20 +13,20 @@ pub struct HeadMotion {
 
 #[context]
 pub struct CreationContext {
-    pub center_head_position: Parameter<HeadJoints, "control.center_head_position">,
-    pub inner_maximum_pitch: Parameter<f32, "control.head_motion.inner_maximum_pitch">,
-    pub maximum_velocity: Parameter<HeadJoints, "control.head_motion.maximum_velocity">,
-    pub outer_maximum_pitch: Parameter<f32, "control.head_motion.outer_maximum_pitch">,
-    pub outer_yaw: Parameter<f32, "control.head_motion.outer_yaw">,
+    pub center_head_position: Parameter<HeadJoints, "center_head_position">,
+    pub inner_maximum_pitch: Parameter<f32, "head_motion.inner_maximum_pitch">,
+    pub maximum_velocity: Parameter<HeadJoints, "head_motion.maximum_velocity">,
+    pub outer_maximum_pitch: Parameter<f32, "head_motion.outer_maximum_pitch">,
+    pub outer_yaw: Parameter<f32, "head_motion.outer_yaw">,
 }
 
 #[context]
 pub struct CycleContext {
-    pub center_head_position: Parameter<HeadJoints, "control.center_head_position">,
-    pub inner_maximum_pitch: Parameter<f32, "control.head_motion.inner_maximum_pitch">,
-    pub maximum_velocity: Parameter<HeadJoints, "control.head_motion.maximum_velocity">,
-    pub outer_maximum_pitch: Parameter<f32, "control.head_motion.outer_maximum_pitch">,
-    pub outer_yaw: Parameter<f32, "control.head_motion.outer_yaw">,
+    pub center_head_position: Parameter<HeadJoints, "center_head_position">,
+    pub inner_maximum_pitch: Parameter<f32, "head_motion.inner_maximum_pitch">,
+    pub maximum_velocity: Parameter<HeadJoints, "head_motion.maximum_velocity">,
+    pub outer_maximum_pitch: Parameter<f32, "head_motion.outer_maximum_pitch">,
+    pub outer_yaw: Parameter<f32, "head_motion.outer_yaw">,
 
     pub look_around: Input<HeadJoints, "look_around">,
     pub look_at: Input<HeadJoints, "look_at">,
