+++ crates/control/src/motion/step_planner.rs
@@ -8,12 +8,12 @@ pub struct StepPlanner {}
 
 #[context]
 pub struct CreationContext {
-    pub injected_step: Parameter<Option<Step>, "control.step_planner.injected_step?">,
-    pub inside_turn_ratio: Parameter<f32, "control.step_planner.inside_turn_ratio">,
-    pub max_step_size: Parameter<Step, "control.step_planner.max_step_size">,
-    pub max_step_size_backwards: Parameter<f32, "control.step_planner.max_step_size_backwards">,
-    pub rotation_exponent: Parameter<f32, "control.step_planner.rotation_exponent">,
-    pub translation_exponent: Parameter<f32, "control.step_planner.translation_exponent">,
+    pub injected_step: Parameter<Option<Step>, "step_planner.injected_step?">,
+    pub inside_turn_ratio: Parameter<f32, "step_planner.inside_turn_ratio">,
+    pub max_step_size: Parameter<Step, "step_planner.max_step_size">,
+    pub max_step_size_backwards: Parameter<f32, "step_planner.max_step_size_backwards">,
+    pub rotation_exponent: Parameter<f32, "step_planner.rotation_exponent">,
+    pub translation_exponent: Parameter<f32, "step_planner.translation_exponent">,
 
     pub walk_return_offset: PersistentState<Step, "walk_return_offset">,
 }
@@ -24,12 +24,12 @@ pub struct CycleContext {
     pub sensor_data: Input<SensorData, "sensor_data">,
     pub support_foot: Input<SupportFoot, "support_foot">,
 
-    pub injected_step: Parameter<Option<Step>, "control.step_planner.injected_step?">,
-    pub inside_turn_ratio: Parameter<f32, "control.step_planner.inside_turn_ratio">,
-    pub max_step_size: Parameter<Step, "control.step_planner.max_step_size">,
-    pub max_step_size_backwards: Parameter<f32, "control.step_planner.max_step_size_backwards">,
-    pub rotation_exponent: Parameter<f32, "control.step_planner.rotation_exponent">,
-    pub translation_exponent: Parameter<f32, "control.step_planner.translation_exponent">,
+    pub injected_step: Parameter<Option<Step>, "step_planner.injected_step?">,
+    pub inside_turn_ratio: Parameter<f32, "step_planner.inside_turn_ratio">,
+    pub max_step_size: Parameter<Step, "step_planner.max_step_size">,
+    pub max_step_size_backwards: Parameter<f32, "step_planner.max_step_size_backwards">,
+    pub rotation_exponent: Parameter<f32, "step_planner.rotation_exponent">,
+    pub translation_exponent: Parameter<f32, "step_planner.translation_exponent">,
 
     pub walk_return_offset: PersistentState<Step, "walk_return_offset">,
 }
