+++ crates/control/src/obstacle_filter.rs
@@ -18,8 +18,7 @@ pub struct ObstacleFilter {
 #[context]
 pub struct CreationContext {
     pub field_dimensions: Parameter<FieldDimensions, "field_dimensions">,
-    pub obstacle_filter_configuration:
-        Parameter<ObstacleFilterConfiguration, "control.obstacle_filter">,
+    pub obstacle_filter_configuration: Parameter<ObstacleFilterConfiguration, "obstacle_filter">,
 }
 
 #[context]
@@ -37,15 +36,13 @@ pub struct CycleContext {
     pub cycle_time: Input<CycleTime, "cycle_time">,
 
     pub field_dimensions: Parameter<FieldDimensions, "field_dimensions">,
-    pub goal_post_obstacle_radius:
-        Parameter<f32, "control.obstacle_filter.goal_post_obstacle_radius">,
-    pub obstacle_filter_configuration:
-        Parameter<ObstacleFilterConfiguration, "control.obstacle_filter">,
+    pub goal_post_obstacle_radius: Parameter<f32, "obstacle_filter.goal_post_obstacle_radius">,
+    pub obstacle_filter_configuration: Parameter<ObstacleFilterConfiguration, "obstacle_filter">,
     pub robot_obstacle_radius_at_foot_height:
-        Parameter<f32, "control.obstacle_filter.robot_obstacle_radius_at_foot_height">,
+        Parameter<f32, "obstacle_filter.robot_obstacle_radius_at_foot_height">,
     pub robot_obstacle_radius_at_hip_height:
-        Parameter<f32, "control.obstacle_filter.robot_obstacle_radius_at_hip_height">,
-    pub unknown_obstacle_radius: Parameter<f32, "control.obstacle_filter.unknown_obstacle_radius">,
+        Parameter<f32, "obstacle_filter.robot_obstacle_radius_at_hip_height">,
+    pub unknown_obstacle_radius: Parameter<f32, "obstacle_filter.unknown_obstacle_radius">,
 
     pub detected_robots_bottom:
         PerceptionInput<Option<DetectedRobots>, "VisionBottom", "detected_robots?">,
