+++ crates/control/src/role_assignment.rs
@@ -22,7 +22,7 @@ pub struct RoleAssignment {
 #[context]
 pub struct CreationContext {
     pub field_dimensions: Parameter<FieldDimensions, "field_dimensions">,
-    pub forced_role: Parameter<Option<Role>, "control.role_assignment.forced_role?">,
+    pub forced_role: Parameter<Option<Role>, "role_assignment.forced_role?">,
     pub player_number: Parameter<PlayerNumber, "player_number">,
     pub spl_network: Parameter<SplNetwork, "spl_network">,
 }
@@ -38,7 +38,7 @@ pub struct CycleContext {
     pub cycle_time: Input<CycleTime, "cycle_time">,
 
     pub field_dimensions: Parameter<FieldDimensions, "field_dimensions">,
-    pub forced_role: Parameter<Option<Role>, "control.role_assignment.forced_role?">,
+    pub forced_role: Parameter<Option<Role>, "role_assignment.forced_role?">,
     pub player_number: Parameter<PlayerNumber, "player_number">,
     pub spl_network: Parameter<SplNetwork, "spl_network">,
     pub spl_message: PerceptionInput<IncomingMessage, "SplNetwork", "message">,
