+++ crates/control/src/sonar_filter.rs
@@ -12,27 +12,23 @@ pub struct SonarFilter {
 
 #[context]
 pub struct CreationContext {
-    pub low_pass_filter_coefficient:
-        Parameter<f32, "control.sonar_filter.low_pass_filter_coefficient">,
-    pub maximal_detectable_distance:
-        Parameter<f32, "control.sonar_filter.maximal_detectable_distance">,
-    pub maximal_reliable_distance: Parameter<f32, "control.sonar_filter.maximal_reliable_distance">,
-    pub minimal_reliable_distance: Parameter<f32, "control.sonar_filter.minimal_reliable_distance">,
-    pub sensor_angle: Parameter<f32, "control.sonar_obstacle.sensor_angle">,
+    pub low_pass_filter_coefficient: Parameter<f32, "sonar_filter.low_pass_filter_coefficient">,
+    pub maximal_detectable_distance: Parameter<f32, "sonar_filter.maximal_detectable_distance">,
+    pub maximal_reliable_distance: Parameter<f32, "sonar_filter.maximal_reliable_distance">,
+    pub minimal_reliable_distance: Parameter<f32, "sonar_filter.minimal_reliable_distance">,
+    pub sensor_angle: Parameter<f32, "sonar_obstacle.sensor_angle">,
 }
 
 #[context]
 pub struct CycleContext {
     pub sonar_values: AdditionalOutput<SonarValues, "sonar_values">,
 
-    pub low_pass_filter_coefficient:
-        Parameter<f32, "control.sonar_filter.low_pass_filter_coefficient">,
-    pub maximal_detectable_distance:
-        Parameter<f32, "control.sonar_filter.maximal_detectable_distance">,
-    pub maximal_reliable_distance: Parameter<f32, "control.sonar_filter.maximal_reliable_distance">,
-    pub minimal_reliable_distance: Parameter<f32, "control.sonar_filter.minimal_reliable_distance">,
-    pub middle_merge_threshold: Parameter<f32, "control.sonar_filter.middle_merge_threshold">,
-    pub sensor_angle: Parameter<f32, "control.sonar_obstacle.sensor_angle">,
+    pub low_pass_filter_coefficient: Parameter<f32, "sonar_filter.low_pass_filter_coefficient">,
+    pub maximal_detectable_distance: Parameter<f32, "sonar_filter.maximal_detectable_distance">,
+    pub maximal_reliable_distance: Parameter<f32, "sonar_filter.maximal_reliable_distance">,
+    pub minimal_reliable_distance: Parameter<f32, "sonar_filter.minimal_reliable_distance">,
+    pub middle_merge_threshold: Parameter<f32, "sonar_filter.middle_merge_threshold">,
+    pub sensor_angle: Parameter<f32, "sonar_obstacle.sensor_angle">,
 
     pub fall_state: Input<FallState, "fall_state">,
     pub sensor_data: Input<SensorData, "sensor_data">,
