+++ crates/control/src/whistle_filter.rs
@@ -13,8 +13,8 @@ pub struct WhistleFilter {
 
 #[context]
 pub struct CreationContext {
-    pub buffer_length: Parameter<usize, "control.whistle_filter.buffer_length">,
-    pub minimum_detections: Parameter<usize, "control.whistle_filter.minimum_detections">,
+    pub buffer_length: Parameter<usize, "whistle_filter.buffer_length">,
+    pub minimum_detections: Parameter<usize, "whistle_filter.minimum_detections">,
 }
 
 #[context]
@@ -22,8 +22,8 @@ pub struct CycleContext {
     pub sensor_data: Input<SensorData, "sensor_data">,
     pub cycle_time: Input<CycleTime, "cycle_time">,
 
-    pub buffer_length: Parameter<usize, "control.whistle_filter.buffer_length">,
-    pub minimum_detections: Parameter<usize, "control.whistle_filter.minimum_detections">,
+    pub buffer_length: Parameter<usize, "whistle_filter.buffer_length">,
+    pub minimum_detections: Parameter<usize, "whistle_filter.minimum_detections">,
     pub detected_whistle: PerceptionInput<Whistle, "Audio", "detected_whistle">,
 }
