+++ crates/vision/Cargo.toml
@@ -5,9 +5,14 @@ name = "vision"
 version = "0.1.0"
 
 [dependencies]
+approx = { workspace = true }
 color-eyre = { workspace = true }
+compiled-nn = { workspace = true }
 context_attribute = { workspace = true }
+itertools = { workspace = true }
 filtering = { workspace = true }
 framework = { workspace = true }
 nalgebra = { workspace = true }
+ordered-float = { workspace = true }
+rand = { workspace = true }
 types = { workspace = true }
