+++ crates/vision/src/camera_matrix_extractor.rs
@@ -3,6 +3,8 @@ use context_attribute::context;
 use framework::MainOutput;
 use types::{hardware::Image, CameraMatrices, CameraMatrix};
 
+use crate::CyclerInstance;
+
 pub struct CameraMatrixExtractor {}
 
 #[context]
@@ -12,6 +14,7 @@ pub struct CreationContext {}
 pub struct CycleContext {
     pub camera_matrices: RequiredInput<Option<CameraMatrices>, "Control", "camera_matrices?">,
     pub image: Input<Image, "image">, // required for correct module order
+    pub instance: CyclerInstance,
 }
 
 #[context]
@@ -25,7 +28,14 @@ impl CameraMatrixExtractor {
         Ok(Self {})
     }
 
-    pub fn cycle(&mut self, _context: CycleContext) -> Result<MainOutputs> {
-        Ok(MainOutputs::default())
+    pub fn cycle(&mut self, context: CycleContext) -> Result<MainOutputs> {
+        let camera_matrix = match context.instance {
+            CyclerInstance::VisionTop => &context.camera_matrices.top,
+            CyclerInstance::VisionBottom => &context.camera_matrices.bottom,
+        };
+
+        Ok(MainOutputs {
+            camera_matrix: Some(camera_matrix.clone()).into(),
+        })
     }
 }
