+++ crates/vision/src/image_receiver.rs
@@ -3,7 +3,7 @@ use context_attribute::context;
 use framework::MainOutput;
 use types::{
     hardware::{Image, Interface},
-    CameraPosition, Rgb, YCbCr444,
+    CameraPosition,
 };
 
 use crate::CyclerInstance;
@@ -22,7 +22,6 @@ pub struct CycleContext {
 #[context]
 pub struct MainOutputs {
     pub image: MainOutput<Image>,
-    pub average_color: MainOutput<Rgb>,
 }
 
 impl ImageReceiver {
@@ -37,41 +36,8 @@ impl ImageReceiver {
                 CyclerInstance::VisionTop => CameraPosition::Top,
                 CyclerInstance::VisionBottom => CameraPosition::Bottom,
             })?;
-        let mut red = 0;
-        let mut green = 0;
-        let mut blue = 0;
-        for y in 0..image.height {
-            for x in 0..image.width_422 {
-                let pixel = image.buffer[(y * image.width_422 + x) as usize];
-                let left_color: Rgb = YCbCr444 {
-                    y: pixel.y1,
-                    cb: pixel.cb,
-                    cr: pixel.cr,
-                }
-                .into();
-                red += left_color.r as u32;
-                green += left_color.g as u32;
-                blue += left_color.b as u32;
-                let right_color: Rgb = YCbCr444 {
-                    y: pixel.y2,
-                    cb: pixel.cb,
-                    cr: pixel.cr,
-                }
-                .into();
-                red += right_color.r as u32;
-                green += right_color.g as u32;
-                blue += right_color.b as u32;
-            }
-        }
-        let amount_of_additions = 2 * image.buffer.len();
         Ok(MainOutputs {
             image: image.into(),
-            average_color: Rgb::new(
-                (red / amount_of_additions as u32) as u8,
-                (green / amount_of_additions as u32) as u8,
-                (blue / amount_of_additions as u32) as u8,
-            )
-            .into(),
         })
     }
 }
