+++ crates/vision/src/lib.rs
@@ -6,6 +6,7 @@ pub mod image_receiver;
 pub mod image_segmenter;
 pub mod line_detection;
 pub mod perspective_grid_candidates_provider;
+mod ransac;
 pub mod robot_detection;
 pub mod segment_filter;
